Fix intermitent test failure on index page feature spec
Since we are using faker to generate data, it sometimes generate
the same job title for both new job and old job, and we rely on this
two titles to check if the test is valid or not and it fails.

This commit adds a static text to the old job created in the spec and
fix the intermitent failure.

Closes #92