Upgrade Rails version to 5.2.1
This commit upgrades the rails version to have the most recent
version of the framework and all the security patches.

After upgraded, the task: `app:update` was executed, generating and
changing old configs to the new pattern. This also changes the
config/application.rb to load the default configs for Rails 5.2 (instead
of manually activating a flag through an initializer file)