Use SQL instead of Active Record's schema dumper
When someone try to setup our project using bin/rails db:schema:load (to
do a faster load of the database instead of running all migrations), or
the developer tries to run the test with the recently created database
it makes the database state wrong because all of the text search
indexes aren't being created. This occurs because schema.rb can't have
specific databases features.