Use Uglifier with harmony flag to uglify ES6
In order to use ES6 scripts present on app/assets, we need to set the
Uglifier with the harmony flag. This is not the ideal way. The best way
is to use Webpacker in the future.

TODO: Move all ES6 scripts from app/javascripts to app/packs to use
Webpacker.