Handle company custom subdomain
If the request contains a non-empty subdomain, try to find a company
with the corresponding custom domain. If not found, raises an
appropriate error response. If found, stores the company record in the
controller instance so it is available to be used by views and
controller actions.

If requesting the vacancies list (index page) and there is a custom
domain, use the company to filter the vacancies list.