Handle company custom subdomain (#147)
* Handle company custom subdomain

If the request contains a non-empty subdomain, try to find a company
with the corresponding custom domain. If not found, raises an
appropriate error response. If found, stores the company record in the
controller instance so it is available to be used by views and
controller actions.

If requesting the vacancies list (index page) and there is a custom
domain, use the company to filter the vacancies list.

* Add test case for searching vacancies filtering by a specific company

* Add feature spec for accessing index with a company specific subdomain

* Refactor Searchs.Vacancy#list method

The new version does the same thing but has a more concise code.

* Refactor ApplicationController

Applying @fpgentil's code readability enhancement suggestions.

* Apply minor code style fix

Use double quotes when the string literal contains a single quote.

* Refactor ApplicationController

Move the company subdomain check from the `filter_by_company` method to
the `if parameter` of the `before_action` call.

* Update brakeman ignore entry