Add SECRET_KEY_BASE env to Dockerfile
The SECRET_KEY_BASE environment variables can be set when you run the
image to production. But if you only build the image locally without
manually setting it, it causes an error on the build process. This
commit adds a default SECRET_KEY_BASE env for the Dockerfile that you
can replace it when you run in your environment.