Implement feature to ban users
We currently have some users that cause some spam in the platform and we
need to ban them. We can't only remove they from the platform, because
this can be easily reverted (they can re-signup).

This commit adds a flag in the users table to identify if the user is
banned. If the flag is present, we don't allow them to login.